% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{random_walk}
\alias{random_walk}
\title{Random walk on Z^2}
\usage{
random_walk(n)
}
\arguments{
\item{n}{Integer specifying the number of steps of the generated walk.}
}
\value{
Vector of coordinates for the walk
}
\description{
Generates a random walk in 2-dimensional integer space
}
\examples{
random_walk(n = 10)

}
